Include summary sample sums and counts if present.
This commit finally unlocks the ability for the Prometheus client
users of the Summary metric type to automatically get total counts
and sums partitioned by labels.  It exposes them through two
synthetic variables, if the underlying data is present.

The following is the base case:

    foo_samples{quantile=0.5}  = <?>
    foo_samples{quantile=0.99} = <?>

The following results with this change:

    foo_samples_sum   = <?>
    foo_samples_count = <?>

Change-Id: I75b5ea0d8c851da8c0c82ed9c8ac0890e4238f87