Use spaces instead of tabs to align equals (=)
Using tabs to align things has the downside that people who have set the
tabstop to be of a different lenght than the orignial author's run the
risk of not having the alignment stay. In the current Makefile before
this change people with tabstop of 4 do not see the equals line up.