Add GaugeFuncVec collecting vectors of GaugeFuncs
GaugeFunc is great for datasources that can provide instant collection
of metrics, like `sql.DB.Stats()`, however they don't allow multiple
gauges on the same metric name with different label values.

A GaugeFuncVec works similarly to a GaugeVec, but instead of creating
Gauge metrics, it accepts a function as the datasource and registers it
for a given set of labels.

Most of the code was already present in the metricVec, we just need to
skip the metric creation part and provide method for adding metrics
directly.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>