process_collector: fill in virtual and resident memory values on macOS (#1600)
Unfortunately, these values aren't available from getrusage(2), or any other
builtin Go API.  Using cgo is one alternative.  It's possible to conditionalize
everything such that cgo can remain disabled on non-Darwin platforms, or even
when cross-compiling Darwin executables on a non-Darwin platform (and stub in
code that causes the metrics to not be exported).  `CGO_ENABLED=1` is set by
default on macOS, but unfortunately is off for the non-host architecture, even
when gcc supports cross-compiling.  (e.g. building with GOARCH=amd on an M2 mac
skipped the cgo code.)  I think that's too subtle of a distinction to rely on
cgo.

There's no builtin equivalent of `syscall.NewLazyDLL()` and `.NewProc()` on
macOS that Go provides for Windows, so we're stuck with a 3rd party dependency.
But it seems stable, maintained, ang getting a fair amount of usage.  I'm
avoiding their struct deserialization because these native structs are packed
differently than the equivalent Go structs, which was causing bad values to be
returned.

The code is heavy with inline comments, and I tried keeping the type names the
same as the C code to make it easier to search for them.  I'm not sure that we
need to do the `mach_vm_region()` call to adjust the `task_info()` values,
because I've never seen that conditional evaluate to True on either amd64,
arm64, or when amd64 is run under Rosetta.  But this is what `ps(1)` does, and
I think it's reasonable to try to match that unless somebody knows it's dead
code.

Signed-off-by: Matt Harbison <mharbison72@gmail.com>