tests: add coverage for the virtual and resident memory support on macOS
Here we have to use cgo, but since it's conditionalized to 1) tests, 2) only on
Darwin, 3) when `CGO_ENABLED=1`, and 4) gracefully skips when it is disabled, I
think it's OK.  There were many layers to the C types that are defined, and
subtle sizing and offset issues that the C compiler would handle that need to be
correct here, so I think it's worth having around.  I don't expect any of the
sizes or offsets to change until a new architecture is supported for macOS, but
these are definitely not the more documented APIs.

Signed-off-by: Matt Harbison <mharbison72@gmail.com>