Add WrapCollectorWith and WrapCollectorWithPrefix (#1766)
I found myself in a situation where I had to instantiate multiple
instances of a third party library that registered some metrics and
never un-registered them.

In my app's lifecycle I needed to un-register them and then register
them again, but I couldn't achieve that by wrapping the Registerer.

I decided to register that library's metrics in a separate Registry and
register that Registry, but in order to handle multiple instances of the
library, I also needed to wrap it with labels, and while such
functionality existed already as part of the Registerer wrapping, it
wasn't exposed.

This PR just exposes Collector wrapping as a public function.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>