This change enables setting min and max values for gauges. It prevents the gauge's value to become smaller/greater than the specified limits. The values are optional and don't change the gauge's behaviour if not given.
Example use case:
A log stream increases or decreases a user session gauge if a corresponding login/logoff log entry is seen. Specifying a minimum value of 0 for the gauge helps to prevent the user session gauge to get negative if the very first message that triggers the gauge is a logoff message. This might happen if the user is already logged in and the prometheus service is restarted, which resets the gauge values to 0.

Signed-off-by: Michael Jenny <prinzdezibel@gmail.com>